fix：电销单条问题修复

+++ marketing-cs/src/main/java/com/br/marketing/mapper/PhoneSaleExtendShuheMapperBase.java
@@ -2,12 +2,11 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.PhoneSaleExtendShuhe;
 import com.br.marketing.entity.PhoneSaleExtendShuheExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-import java.util.List;
-
 public interface PhoneSaleExtendShuheMapperBase {
-    int countByExample(PhoneSaleExtendShuheExample example);
+    long countByExample(PhoneSaleExtendShuheExample example);
 
     int deleteByExample(PhoneSaleExtendShuheExample example);
